stock-trade-producer/requirements.txt
websocket-client==1.8.0
confluent-kafka==2.4.0